chore(sidecar): rename mevboost > constraint_client
