+++ bolt-sidecar/src/client/mod.rs
@@ -1,5 +1,5 @@
 pub mod commit_boost;
-pub mod mevboost;
+pub mod constraint_client;
 pub mod pubsub;
 pub mod rpc;
